Add 'grub_default'
usage: grub_default [--force] (release|restore)

Facilitates the setting of the default grub menu entry to the [kernel]
file in /boot containing $(uname -r).
The aim is to enable the automatic restoration from hibernation when a
non-default kernel was used. This is needed as if there is a version
mismatch between the hibernation image in swap and the booted kernel,
the image is wiped -- preventing possible issues, but losing the
hibernated session.