Add 'bin/recoll-query', a wrapper around the 'recoll' search.
Inspiration taken from `recollq' -- the command line program that is
installed if the QT GUI is not wanted.
This script takes all the options that could be passed to `recoll[q]'
(as they can be directly passed to `recoll' as well) and always searches
from the current (or specified) directory. Because this makes sense from
the command line. Consequently it emulates the 'ANY' 'ALL' and
'filename' modes presented by the GUI itself in query language, and
makes phrase quoting more concise (if slightly less consistent).