Fix 'recoll-query' and add a 'rq' symlink.
Fix the 'exec recoll' line to not include the query type ($qtype) as it
is emulated by the script.
To complement the 'rq' symlink, add detection so that if the script is
called as 'rq' it only outputs the matching file paths. Done with the
basic (-b) option piped to 'cut -c8-', removing the first 7 characters
of the uri (file://) to give the local path.