Minor lib/*.sh include changes, add notify.sh, rewrite batt.
Change the include logic for library scripts to be slightly simpler for
the ctual source lines. Add a variable check to prevent lib being
sourced more than once.
Add 'lib/notify.sh' as a rewrite of 'lib/send_message.sh', keeping the
interface similar to 'lib/output.sh' with 'info' and 'error' functions,
etc. It uses 'notify-send' if XAUTHORITY, and 'echo's if connected to a
terminal or logging (LOG set). Add a compat layer for send_message
(including symlinking to notify.sh). Change 'sbin/wifi', etc. to use now
interface.
Rewrite 'bin/batt' so it should work with multiple batteries, only
source libraries when needed, and add in a '-n' option for
'notify-send'ing (for use with keyboard shortcuts).
BUGS:
    'colour.sh' seems to be broken