Add getusers - lists all users on the system (file/nis/db)
As this was easist to do as a mini C program, and I don't want to be
commiting binaries, I create a wrapper called gen.sh, which bin/getusers
points to. gen.sh looks in share/src/ for getusers.* (${0##*/}.*) and
tries to compile it as share/src/gen/getusers, followed by `exec'ing it.