Second attempt at auto translation of .hgignore to .gitignore
First attempt was a sed script directly in my post-checkout hook, but it
missed too many edge cases.
This should be re-implemented using the result of sre_parse.parse() so
that the actual parsing used by mercurial (i.e. python regexes) can be
used to translate it, rather than a poor-mans reimplementation.